puya: add support for PY32F002B
The PY32F002B series uses different DBGMCU IDCODE values and there
does not seem to be a register containing RAM and flash size.

While the procedure for preparing flash access (copying parameters
from factory-programmed EPPARAx registers to flash peripheral
registers) is the same, the bit allocation inside the registers is
slightly different.

The structure and values of the DBGMCU IDCODE register are
undocumented but the vendor SDK splits it into DEV_ID and REV_ID
fields. We use the DEV_ID fields of the IDCODE values observed on
PY32F002AW15U and PY32F002BF15P6 to distinguish between the two
families. An internet search shows that at least the PY32F002BW15 uses
the same value as the PY32F002BF15P6. The full IDCODE values are
retained as comments to make it easier to fix the code later if it
turns out the DEV_ID/REV_ID split is incorrect.

For PY32F002A the EPPARA4 PRETPE mask value was kept the same as
before (14 bits). The Reference Manual has conflicting information
about it: EPPARA4 26:16 (which are 11 bits) is supposed to contain
PRETPE 11:0 (which are 12 bits). FLASH_PRETPE, which is documented as
having to be initialised from the EPPARA4 value, has 14 bits for
PRETPE. The official Keil Device Family Pack v1.2.1 from Puya uses the
full upper 16 bit of EPPARA4 so presumably the reserved bits are
simply 0 and using the full 14 bit mask is probably fine.